Dummy Server/requirements.txt
tensorflow==2.17.0
numpy==1.26.4
flask==2.2.5
flask-cors==5.0.1
pillow==11.1.0